requirements.txt
Django>=4.2,<7.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.0
django-filter>=23.0
drf-spectacular>=0.27
python-dotenv>=1.0
psycopg2-binary>=2.9
kafka-python>=2.0
